+++ tests/context.py
@@ -1,6 +1,0 @@
-import os
-import sys
-
-import pyatag
-
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))